requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mypy==0.761
mypy-extensions==0.4.3
nose==1.3.7
numpy==1.18.0
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytube==9.5.3
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22
scipy==1.4.1
six==1.13.0
tabulate==0.8.6
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
